support rewrite
this required reaching in to the block manager suggesting that
the SerializationContext api is insufficient for this use.
It is likely that making a new block manager on write will
solve the issues encountered with rewrite as they revolved
around keeping old blocks that were used on read and reusing
them correctly on write. By making a new block manager on write
no reuse will be needed and the old blocks can be left untouched
referring back to the read file.